+++ components/google-analytics-ga4/src/actions/accounts.ts
@@ -1,7 +1,8 @@
-import { action } from "@prismatic-io/spectral";
+import { action, outputSchema } from "@prismatic-io/spectral";
 import { createAnalyticsClient } from "../client";
 import { listAccountsExamplePayload } from "../examplePayloads";
 import { listAccountsInputs } from "../inputs";
+import { listAccountsOutputSchema } from "../outputSchemas";
 import type { Account } from "../types";
 import { paginateRecords } from "../util";
 const listAccounts = action({
@@ -10,7 +11,12 @@ const listAccounts = action({
     description: "Return a list of accounts accessible by the caller",
   },
   inputs: listAccountsInputs,
-  perform: async (context, { connection, fetchAll, pageSize, pageToken }) => {
+  outputSchema: outputSchema({
+    type: "actionOutput",
+    schema: listAccountsOutputSchema,
+  }),
+  performSafety: "notAllowed",
+  perform: async (context, { connection, fetchAll, pagination }) => {
     const client = createAnalyticsClient({
       connection,
       endpointType: "adminv1beta",
@@ -20,14 +26,19 @@ const listAccounts = action({
       client,
       "/accounts",
       {
-        pageSize,
-        pageToken,
+        pageSize: pagination.pageSize,
+        pageToken: pagination.pageToken,
       },
       fetchAll,
       "accounts",
     );
     return { data };
   },
+  examplePerform: async (): Promise<{
+    data: unknown;
+  }> => ({
+    ...listAccountsExamplePayload,
+  }),
   examplePayload: listAccountsExamplePayload,
 });
 export default { listAccounts };